slintel_server/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
python-resize-image==1.1.19
